requirements.txt
streamlit==1.41.1
google-generativeai==0.8.3
selenium==4.28.1
beautifulsoup4==4.13.3